+++ .github/workflows/nodejs.yml
@@ -4,7 +4,7 @@ on:
   push:
     branches:
       - master
-      - env-list
+      - http-client
 
 jobs:
   build:
